Feature/add rabbitmq (#7)
* adding-attributes-and-line-breaks

* adding-sorting

* adding-users-queue-configuration

* adding-accounts-queue-configuration

* changing-controllers

* adding-operations-queue-configuration

* restructuring-consumer-logic

* changing-auth-structure

* changing-consumer-structure-in-accounts

* fixing-get-method-and-removing-unnecessary

* correction

* cleaning

* returning-controllers-to-their-previous-state

* updating-projects-settings

* updating-messages-logic

* setting-up-sending-user-data

* setting-up-receiving-user-data

* setting-up-account-data-sharing

* changing_the_update_function

* changing_the_error_message

* adding-consts-for-rabbit

+++ Microservices/AccountsMicroservice/AccountsBLL/Services/Implementations/FinancialAccountService.cs
@@ -1,5 +1,7 @@
 ﻿using Abp.Domain.Entities;
+using Accounts.BusinessLogic.Consumers;
 using Accounts.BusinessLogic.Models;
+using Accounts.BusinessLogic.Producers;
 using Accounts.BusinessLogic.Services.Interfaces;
 using Accounts.DataAccess.Entities;
 using Accounts.DataAccess.Settings;
@@ -12,28 +14,47 @@ namespace Accounts.BusinessLogic.Services.Implementations
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
+        private readonly IMessageConsumer _consumer;
+        private readonly IMessageProducer _producer;
 
-        public FinancialAccountService(IUnitOfWork unitOfWork, IMapper mapper)
+        public FinancialAccountService(IUnitOfWork unitOfWork, IMapper mapper, 
+            IMessageConsumer consumer, IMessageProducer producer)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
+            _consumer = consumer;
+            _producer = producer;
         }
 
         public async Task AddAsync(FinancialAccountModel addModel, CancellationToken cancellationToken)
         {
+            int userId = _consumer.ConsumeMessage(addModel.UserId);
+
+            if (userId == 0)
+            {
+                throw new Exception("The user is not logged in");
+            }
+
             var account = _mapper.Map<FinancialAccount>(addModel);
             await _unitOfWork.FinancialAccounts.AddAsync(account, cancellationToken);
 
             await _unitOfWork.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(int id, CancellationToken cancellationToken)
+        public async Task DeleteAsync(int userId, int id, CancellationToken cancellationToken)
         {
             var account = await _unitOfWork.FinancialAccounts.GetByIdAsync(id, cancellationToken);
 
             if (account == null)
             {
                 throw new EntityNotFoundException("Account not found");
+            }
+
+            int receivedUserId = _consumer.ConsumeMessage(account.UserId);
+
+            if (receivedUserId == 0 || userId != receivedUserId)
+            {
+                throw new Exception("The user is not logged in");
             }
 
             await _unitOfWork.FinancialAccounts.DeleteAsync(id, cancellationToken);
@@ -50,12 +71,21 @@ namespace Accounts.BusinessLogic.Services.Implementations
             return accountsList;
         }
 
-        public async Task<List<FinancialAccountModel>> GetAccountsByUserIdAsync(int userId, 
+        public async Task<List<FinancialAccountModel>> GetAccountsByUserIdAsync(int userId,
             PaginationSettings paginationSettings, CancellationToken cancellationToken)
         {
+            int id = _consumer.ConsumeMessage(userId);
+
+            if (id == 0)
+            {
+                throw new Exception("The user is not logged in");
+            }
+
             var accounts = await _unitOfWork.FinancialAccounts
-                .GetAccountsByUserIdAsync(userId, paginationSettings, cancellationToken);
+                .GetAccountsByUserIdAsync(id, paginationSettings, cancellationToken);
             var accountsList = _mapper.Map<List<FinancialAccountModel>>(accounts);
+
+            _producer.SendMessages(accountsList);   
 
             return accountsList;
         }
@@ -74,13 +104,21 @@ namespace Accounts.BusinessLogic.Services.Implementations
             return accountModel;
         }
 
-        public async Task UpdateAsync(int id, FinancialAccountModel updateModel, CancellationToken cancellationToken)
+        public async Task UpdateAsync(int userId, int id, FinancialAccountModel updateModel, 
+            CancellationToken cancellationToken)
         {
             var account = await _unitOfWork.FinancialAccounts.GetByIdAsync(id, cancellationToken);
 
             if (account == null)
             {
                 throw new EntityNotFoundException("Account not found");
+            }
+
+            int receivedUserId = _consumer.ConsumeMessage(account.UserId);
+
+            if (receivedUserId == 0 || userId != receivedUserId)
+            {
+                throw new Exception("The user is not logged in");
             }
 
             var updateAccount = _mapper.Map<FinancialAccount>(updateModel);
